requirements.txt
alembic==1.6.4
banal==1.0.6
dataset==1.5.0
greenlet==1.1.0
Mako==1.1.4
MarkupSafe==2.0.1
mysqlclient==2.0.3
numpy==1.20.3
Pillow==8.2.0
Pypubsub==4.0.3
python-dateutil==2.8.1
python-editor==1.0.4
six==1.16.0
SQLAlchemy==1.4.15
wxPython==4.1.1
